+++ header.php
@@ -87,6 +87,14 @@
 
 		<?php wp_head(); ?>
 
+		<!-- Salesforce Contact For ReCaptcha -->
+		<script src="https://www.google.com/recaptcha/api.js"></script>
+		<script>
+		 function timestamp() { var response = document.getElementById("g-recaptcha-response"); if (response == null || response.value.trim() == "") {var elems = JSON.parse(document.getElementsByName("captcha_settings")[0].value);elems["ts"] = JSON.stringify(new Date().getTime());document.getElementsByName("captcha_settings")[0].value = JSON.stringify(elems); } } setInterval(timestamp, 500); 
+		</script>
+		<!-- end Salesforce Contact For ReCaptcha -->
+
+
 	</head>
 
 	<body <?php body_class(); ?> itemscope itemtype="http://schema.org/WebPage">